Adding a stop file.

+++ main.sh
@@ -102,17 +102,19 @@ date > /home/ubuntu/foo.txt
 export NS3_HOME=/home/ubuntu/workspace/source/ns-3.26
 while [  $COUNTER -lt $TS_CYCLES ]; do
 
-    cd /home/ubuntu/tap
-	DATENOW=$(date +"%y_%m_%d_%H_%M")
-	mkdir -p var/archive/$DATENOW
-	mv var/log/* var/archive/$DATENOW/
+    if [ ! -f /home/ubuntu/stop.txt ]; then
+        cd /home/ubuntu/tap
+        DATENOW=$(date +"%y_%m_%d_%H_%M")
+        mkdir -p var/archive/$DATENOW
+        mv var/log/* var/archive/$DATENOW/
 
-	echo "python3 main.py -n $TS_NODES -t $TS_TIME -to $TS_TIMEOUT -s $TS_SIZE -ns $TS_SPEED -np $TS_PAUSE -c $COUNTER full"
-	python3 main.py -n $TS_NODES -t $TS_TIME -to $TS_TIMEOUT -s $TS_SIZE -ns $TS_SPEED -np $TS_PAUSE -c $COUNTER full
+        echo "python3 main.py -n $TS_NODES -t $TS_TIME -to $TS_TIMEOUT -s $TS_SIZE -ns $TS_SPEED -np $TS_PAUSE -c $COUNTER full"
+        python3 main.py -n $TS_NODES -t $TS_TIME -to $TS_TIMEOUT -s $TS_SIZE -ns $TS_SPEED -np $TS_PAUSE -c $COUNTER full
 
-	cd /home/ubuntu/EC2TestAutomator
-	echo "python3 statscollector2.py -ns $TS_SPEED -np $TS_PAUSE $TS_NAME $TS_NODES $TS_TIME $TS_TIMEOUT $TS_SIZE"
-	python3 statscollector2.py -ns $TS_SPEED -np $TS_PAUSE $TS_NAME $TS_NODES $TS_TIME $TS_TIMEOUT $TS_SIZE
+        cd /home/ubuntu/EC2TestAutomator
+        echo "python3 statscollector2.py -ns $TS_SPEED -np $TS_PAUSE $TS_NAME $TS_NODES $TS_TIME $TS_TIMEOUT $TS_SIZE"
+        python3 statscollector2.py -ns $TS_SPEED -np $TS_PAUSE $TS_NAME $TS_NODES $TS_TIME $TS_TIMEOUT $TS_SIZE
+    fi
 
 	let COUNTER=COUNTER+1
 	echo $COUNTER >> /home/ubuntu/foo.txt
